Issue #45 Market Subscription Standardization
When subscribing to markets the callback will now receive something that
looks like this:

[  "REMOVED ID1", "REMOVED ID2", ...,
    { "id": "UPDATED ID3", data... },
    { "id": "UPDATED ID4", data... },
    ...
    [ FILL_OPERATION_TYPE_ID, { fill order operation data } ],
    [ FILL_OPERATION_TYPE_ID, { fill order operation data } ],
    [ FILL_OPERATION_TYPE_ID, { fill order operation data } ],
    [ FILL_OPERATION_TYPE_ID, { fill order operation data } ]
]

When ever an order is removed from the order book, its ID will be
included as a string.  When an order is modified the new order data is included as
an object. And the operations describing how orders were matched and
what fees are paid will be included as an operation, aka array.

Also added means to unsubscribe from full account data (issue #166)