added function to database public api get_account_history_public
requested in: https://github.com/bitshares/bitshares-core/issues/105
i tried first to move the original get_account_history function from the
history api to the database api but i had errors in the wallet.cpp so i
decided to make a new call. not sure if this is the way to go or either
the name but function is there and it works just as get_account_history.