[config.ini] Add RAM-usage optimization to newly created config.ini files
The reasoning here is that maintainers and operators don't need to
customize the configuration file - if they never used bitshares-core,
while it keeps integrity if a config.ini already exists - in order to
have the BitShares backend running in low memory mode. The number 1000
should be sufficient for most operations (even exchanges) to include all
needed operations in the exchanges' account.