Swap logic: if config file does not exist, create one and parse it
Previously, a newly written config file was ignored.
For the included plugins that didn't matter though, because all config file
options (and their default values) are also CLI options, and CLI defaults
always apply, no matter if a config file exists or not.

Now, a new config file is written with the default option values (if any),
and parsed immediately after that. In other words, the config defaults
are applied again. This does not result in a different outcome, at least
with the included plugins.