set default quote asset_id to 1.3.1
In current help message of create_asset, printed in gethelp, gives an
example of option with quote asset_id of "1.3.0". This asset id is
not a valid one, since the code requires the id to be "1.3.1" on
creating a new asset.

This change set the default quote asset_id to "1.3.1", so that the
gethelp message would give a valid option example.