BSIP 40: Error reporting, part 2
Add the wrapping to report errors back to the user... such as it is
Eventually, I'd like to add more logic towards generating a user-
comprehensible error message, but this will suffice for now.

This implementation provides enough information to construct, in
conjunction with the restriction list and operation in question, a
message describing the error in detail to the user... but actually
assembling that message is, for now, regarded as a UI problem.